Various cleanups found by FindBugs http://findbugs.sourceforge.net
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@9101 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/sphinx4/edu/cmu/sphinx/frontend/test/DataBlockerTest.java
@@ -113,11 +113,11 @@ public class DataBlockerTest extends BaseDataProcessor {
 
 
     /**
-     * Tests wether the samples of all <code>Data</code>s in the list are ordered in increasing order with +1
+     * Tests whether the samples of all <code>Data</code>s in the list are ordered in increasing order with +1
      * increments.
      */
     public static boolean hasIncreasingOrder(List<Data> output, int lastValue) {
-        double dataCounter = 0;
+        int dataCounter = 0;
 
         for (Data data : output) {
             if (data instanceof DoubleData) {
